Check tenant when reusing public net

+++ taf/testlib/virtual_env.py
@@ -28,7 +28,9 @@ limitations under the License.
 #  "instance_type": "settings",
 #  "id": "993",
 #  "images_share_path": "/mnt/berta/oses/openstack",
-#  "mgmt_ip_cidr": "11.212.23.11/8",
+#  "external_net_ip_cidr": "11.212.23.0/24",
+#  "external_net_pool_start": "100",
+#  "external_net_pool_end": "199",
 #  "other_configs": {"ovs_type": "ovs",}
 # }
 ###############################################################################
@@ -107,7 +109,7 @@ class VirtualEnv(object):
         self.opts = opts
         self.settings = self._get_settings(self.opts.env)
         self.tempest_path = self.opts.tempest_path
-        self.reuse_venv = self.opts.reuse_venv
+        self.reuse_venv = eval(self.opts.reuse_venv)
         self.neutron_extensions = None
 
         import tempest
@@ -248,7 +250,7 @@ class VirtualEnv(object):
         from tempest.common import waiters
         return waiters.wait_for_server_status(self.handle.servers_client, vm_id, status)
 
-    def ensure_public_access(self, try_reuse=False, networks_client=None, routers_client=None,
+    def ensure_public_access(self, try_reuse=True, networks_client=None, routers_client=None,
                              name=None, tenant_id=None):
         """Create or reuse public/external router & network.
 
@@ -271,12 +273,17 @@ class VirtualEnv(object):
             tenant_id = self.tenant_id
 
         _net_cfg = self.config.network
-        _mgmt_ip_cidr = self.settings.get('mgmt_ip_cidr')
-        assert _mgmt_ip_cidr
-        net_ip = netaddr.IPNetwork(_mgmt_ip_cidr)
+        _external_net_ip_cidr = self.settings.get('external_net_ip_cidr')
+        assert _external_net_ip_cidr
+        net_ip = netaddr.IPNetwork(_external_net_ip_cidr)
+
+        _external_net_pool_start = self.settings.get('external_net_pool_start')
+        assert _external_net_pool_start
+        _external_net_pool_end = self.settings.get('external_net_pool_end')
+        assert _external_net_pool_end
 
         # Try to reuse existing stuff that meets requirements, if desirable (devstack)
-        if try_reuse and self._reuse_public_access(net_ip, routers_client=routers_client):
+        if try_reuse and self._reuse_public_access(net_ip, routers_client=routers_client, tenant_id=tenant_id):
             self.class_logger.debug('Reused')
             return True
 
@@ -292,21 +299,24 @@ class VirtualEnv(object):
         public_network = self.create_public_network(**public_network_kwargs)
         assert public_network
 
-        subnet_name = self.tempest_lib.common.utils.data_utils.rand_name('tempest-public-subnet')
-        allocation_prefix = _mgmt_ip_cidr.rsplit('.', 1)[0]
-        subnet_kwargs = {
-            'cidr': '{}/{}'.format(net_ip.network, net_ip.prefixlen),
-            'routers_client': routers_client,
-            'name': subnet_name,
-            'tenant_id': tenant_id,
-            'ip_version': 4,
-            'allocation_pools': [{
-                'start': '{}.100'.format(allocation_prefix),
-                'end': '{}.254'.format(allocation_prefix)}],
-            'gateway_ip': net_ip.ip,
-            'enable_dhcp': False
-        }
-        self._create_subnet(public_network['id'], **subnet_kwargs)
+        if not public_network['subnets']:
+            subnet_name = self.tempest_lib.common.utils.data_utils.rand_name('tempest-public-subnet')
+            allocation_prefix = _external_net_ip_cidr.rsplit('.', 1)[0]
+            net_ip_gw = net_ip
+            net_ip_gw.value = net_ip.first + 1
+            subnet_kwargs = {
+                'cidr': '{}'.format(net_ip.cidr),
+                'routers_client': routers_client,
+                'name': subnet_name,
+                'tenant_id': tenant_id,
+                'ip_version': 4,
+                'allocation_pools': [{
+                    'start': '{}.{}'.format(allocation_prefix, _external_net_pool_start),
+                    'end': '{}.{}'.format(allocation_prefix, _external_net_pool_end)}],
+                'gateway_ip': '{}'.format(net_ip_gw.ip),
+                'enable_dhcp': False
+            }
+            self._create_subnet(public_network['id'], **subnet_kwargs)
 
         _net_cfg.public_network_id = public_network['id']
 
@@ -325,25 +335,29 @@ class VirtualEnv(object):
 
         return False
 
-    def _get_external_elements(self, routers_client=None):
+    def _get_external_elements(self, routers_client=None, tenant_id=None):
         if not routers_client:
             routers_client = self.handle.os_adm.routers_client
 
         net_filter = {'router:external': True}
         nets = [net['id'] for net in self.handle._list_networks(**net_filter)]
-        router_2_net_map = {}
+        net_2_router_map = {}
         routers_resp = routers_client.list_routers()
         for router in routers_resp['routers']:
-            ext_gw_info = router.get('external_gateway_info')
-            if ext_gw_info:
-                net_id = ext_gw_info.get('network_id')
-                if net_id and net_id in nets:
-                    router_2_net_map[router['id']] = net_id
+            if tenant_id and tenant_id == router.get('tenant_id'):
+                ext_gw_info = router.get('external_gateway_info')
+                if ext_gw_info:
+                    net_id = ext_gw_info.get('network_id')
+                    if net_id and net_id in nets:
+                        net_2_router_map[net_id] = router['id']
 
-        return router_2_net_map
+        if not net_2_router_map:
+            for nets_id in nets:
+                net_2_router_map[nets_id] = None
+        return net_2_router_map
 
     def _delete_external_elements(self, routers_client=None, networks_client=None,
-                                  ports_client=None):
+                                  ports_client=None, tenant_id=None):
         """
         @brief   Look for the external routers & networks and delete them
         """
@@ -353,17 +367,24 @@ class VirtualEnv(object):
         if not networks_client:
             networks_client = self.handle.os_adm.networks_client
 
-        router_2_net_map = self._get_external_elements(routers_client=routers_client)
-        for router_id, network_id in router_2_net_map.items():
-            self.class_logger.debug('Removing external router: (%s)',
-                                    routers_client.show_router(router_id)['router']['name'])
-            self.delete_router(router_id)
+        net_2_router_map = self._get_external_elements(routers_client=routers_client, tenant_id=tenant_id)
+        for network_id, router_id in net_2_router_map.items():
+            if router_id and tenant_id and tenant_id == routers_client.show_router(router_id)['router']['tenant_id']:
+                self.class_logger.debug('Removing external router: (%s)',
+                                        routers_client.show_router(router_id)['router']['name'])
+                self.delete_router(router_id, routers_client, ports_client)
 
-            self.class_logger.debug('Removing external network: (%s)',
-                                    networks_client.show_network(network_id)['network']['name'])
-            networks_client.delete_network(network_id)
+            if tenant_id and tenant_id == networks_client.show_network(network_id)['network']['tenant_id']:
+                self.class_logger.debug('Removing external network: (%s)',
+                                        networks_client.show_network(network_id)['network']['name'])
+                networks_client.delete_network(network_id)
+            else:
+                self.class_logger.debug('Removing of external network: (%s) is not possible - try to reuse it',
+                                        networks_client.show_network(network_id)['network']['name'])
+                return network_id
+        return None
 
-    def _reuse_public_access(self, mgmt_net, routers_client=None):
+    def _reuse_public_access(self, external_net, routers_client=None, tenant_id=None):
         """
         @brief   Search for the external routers & networks
         """
@@ -371,15 +392,16 @@ class VirtualEnv(object):
             routers_client = self.handle.os_adm.routers_client
 
         _net_cfg = self.config.network
-        router_2_net_map = self._get_external_elements(routers_client=routers_client)
-        if router_2_net_map:
-            for router_id, net_id in router_2_net_map.items():
-                rt_obj = routers_client.show_router(router_id)['router']
-                ext_ips = rt_obj['external_gateway_info']['external_fixed_ips']
-                for ip in ext_ips:
-                    if ip['ip_address'] in mgmt_net:
-                        _net_cfg.public_router_id, _net_cfg.public_network_id = router_id, net_id
-                        return True
+        net_2_router_map = self._get_external_elements(routers_client=routers_client, tenant_id=tenant_id)
+        if net_2_router_map:
+            for net_id, router_id in net_2_router_map.items():
+                if router_id and tenant_id and tenant_id == routers_client.show_router(router_id)['router']['tenant_id']:
+                    rt_obj = routers_client.show_router(router_id)['router']
+                    ext_ips = rt_obj['external_gateway_info']['external_fixed_ips']
+                    for ip in ext_ips:
+                        if ip['ip_address'] in external_net:
+                            _net_cfg.public_router_id, _net_cfg.public_network_id = router_id, net_id
+                            return True
         return False
 
     def create_router(self, routers_client=None, name=None, network_id=None, tenant_id=None,
@@ -577,9 +599,12 @@ class VirtualEnv(object):
             del_external_kwargs = {
                 'routers_client': routers_client,
                 'networks_client': networks_client,
-                'ports_client': ports_client
+                'ports_client': ports_client,
+                'tenant_id': tenant_id,
             }
-            self._delete_external_elements(**del_external_kwargs)
+            external_net_id=self._delete_external_elements(**del_external_kwargs)
+            if external_net_id:
+                return networks_client.show_network(external_net_id)['network']
 
         network_kwargs = {
             'networks_client': networks_client,
@@ -628,7 +653,7 @@ class VirtualEnv(object):
 
         @param network_id: the network to create the subnet in
         @type  network_id: str
-        @param cidr: IP/mask of the mgmt interface, None for project default
+        @param cidr: IP/mask of the external network interface, None for project default
         @type  cidr: str
         @param mask_bits: subnet mask length in bits
         @type  mask_bits: int
